Edited version ofcode scripts and minor changes

+++ week1/code/csvtospace.sh
@@ -7,9 +7,21 @@
 # Arguments: 1 -> comma delimited file
 # Date: Oct 2025
 
+
+if [[ -e "$1" ]]; then
+
 echo "Creating a space delimited version of $1 ..."
 
 cat "$1" | tr -s "," " " > "${1%.csv}_space.txt"
 
 echo "Done!"
+
 exit
+
+else 
+
+echo "Please provide a valid file" 
+
+exit
+
+fi
